TsadriuUtilities.String, TsadriuUtilities.DateTime
author:

DateTime:
Fixed an issue where the ToDateTime() and ToNullableDateTime() methods would not take in consideration the DateStyle provided.

String:
Fixed an issue where the GetBetween method was not returning an empty string when one of the tags were not found.
Fixed an issue where the method GetManyBetween would throw an exception when "\n", "\r" and such white-space characters were used. Now it only throws an exception when either one of the tags is empty "" or null.